+++ test/make/check_so_deps.sh
@@ -49,17 +49,6 @@ function confirm_abi_deps() {
 	fi
 
 	cat << EOF > ${suppression_file}
-[suppress_type]
-	name = spdk_nvme_ctrlr_data
-[suppress_type]
-	type_kind = enum
-	changed_enumerators = SPDK_BDEV_NUM_IO_TYPES
-[suppress_type]
-	name = spdk_env_opts
-[suppress_type]
-	name = spdk_app_opts
-[suppress_type]
-	name = spdk_thread
 EOF
 
 	for object in "$libdir"/libspdk_*.so; do