+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## v21.04: (Upcoming Release)
 
+### nvme
+
+Added `spdk_nvme_qpair_get_optimal_poll_group` function and `qpair_get_optimal_poll_group`
+function pointer in spdk_nvmf_transport_ops structure in order to add the qpair to the most
+suitable polling group.
+
 ## v21.01:
 
 ### bdev